+++ internal/cli/output_contract.go
@@ -87,6 +87,8 @@ jira issue view
 jira issue watchers add
 jira issue watchers list
 jira issue watchers remove
+jira issue worklog add
+jira issue worklog list
 jira link-types
 jira me
 jira planning report